+++ audionimbus/src/effect/ambisonics/binaural.rs
@@ -263,7 +263,6 @@ impl AmbisonicsBinauralEffectParams<'_> {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
     use crate::*;
 
     mod apply {