+++ audionimbus/src/effect/ambisonics/decode.rs
@@ -324,7 +324,6 @@ pub enum Rendering {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
     use crate::*;
 
     mod apply {