+++ audionimbus/src/effect/ambisonics/encode.rs
@@ -269,7 +269,6 @@ impl AmbisonicsEncodeEffectParams {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
     use crate::*;
 
     mod apply {