+++ audionimbus/src/effect/ambisonics/panning.rs
@@ -277,7 +277,6 @@ impl AmbisonicsPanningEffectParams {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
     use crate::*;
 
     mod apply {