+++ audionimbus/src/reconstructor.rs
@@ -268,7 +268,6 @@ impl std::fmt::Display for ReconstructorError {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
     use crate::*;
 
     mod reconstructor {